chore: make formartUlr code very short
