+++ components/dashboard/analysis/AnalysisProgress.tsx
@@ -1,5 +1,6 @@
 "use client";
 import { useAnalysisProgress } from '@/hooks/useAnalysisProgress';
+import { formatUrl } from '@/utils/general';
 import { CheckIcon, Circle, XCircleIcon } from 'lucide-react';
 import React from 'react';
 import { CircularProgressbar, buildStyles } from 'react-circular-progressbar';
@@ -31,7 +32,7 @@ const AnalysisProgress: React.FC<AnalysisProgressProps> = ({
     return (
       <div className="bg-red-50 border border-red-200 rounded-lg p-4">
         <div className="flex items-center">
-          <div className="flex-shrink-0">
+          <div className="shrink-0">
             <svg className="h-5 w-5 text-red-400" viewBox="0 0 20 20" fill="currentColor">
               <path fillRule="evenodd" d="M10 18a8 8 0 100-16 8 8 0 000 16zM8.707 7.293a1 1 0 00-1.414 1.414L8.586 10l-1.293 1.293a1 1 0 101.414 1.414L10 11.414l1.293 1.293a1 1 0 001.414-1.414L11.414 10l1.293-1.293a1 1 0 00-1.414-1.414L10 8.586 8.707 7.293z" clipRule="evenodd" />
             </svg>
@@ -51,15 +52,6 @@ const AnalysisProgress: React.FC<AnalysisProgressProps> = ({
         <p className="text-gray-500">No progress data available</p>
       </div>
     );
-  }
-
-  const formatUrl = (url: string) => {
-    if (url.length <= 28) return url;
-
-    const start = url.slice(0, 10); // first 10 chars
-    const end = url.slice(-5); // last 5 chars
-
-    return `${start}...${end}`; // total = 18 chars
   }
